Organize runtime plugin and context files
