+++ docs/context/plans/TASK-02_magpie_speech_omni_migration.md
