+++ docs/context/plans/TASK-03_diffusion_migration.md
