+++ docs/context/plans/TASK-04_cleanup_delete_legacy.md
