+++ docs/context/plans/TASK-05_plugin_architecture.md
