+++ docs/context/plans/batched-inference-api-executive-brief-2026-04-02.md
