+++ docs/context/reports/FLUX2_PERF_WORKLOG.md
