+++ docs/superpowers/plans/2026-03-30-ffi-kernel-architecture.md
@@ -528,7 +528,7 @@ And conditionally include the module loader:
 
 ```cpp
 #if TRTMC_HAS_TVM_FFI
-#include "plugins/tvm_ffi_module_loader.h"
+#include "runtime/plugins/tvm_ffi/tvm_ffi_module_loader.h"
 #endif
 ```
