+++ src/runtime/plugins/shared/plugin_helpers.cpp
@@ -14,7 +14,7 @@
 #include <utility>
 
 #if TRTMC_HAS_TVM_FFI
-#include "plugins/tvm_ffi_module_loader.h"
+#include "runtime/plugins/tvm_ffi/tvm_ffi_module_loader.h"
 #endif
 
 namespace trtmc {
