+++ src/runtime/plugins/tvm_ffi/tvm_ffi_kernel_creator.cpp
@@ -3,7 +3,7 @@
 
 #if TRTMC_HAS_TRT && TRTMC_HAS_TVM_FFI
 
-#include "plugins/tvm_ffi_kernel_plugin.h"
+#include "runtime/plugins/tvm_ffi/tvm_ffi_kernel_plugin.h"
 
 #include <NvInferRuntime.h>
 #include <cstring>
