+++ tests/cpp/test_tvm_ffi_module_loader.cpp
@@ -22,7 +22,7 @@
 
 #if TRTMC_HAS_TRT && TRTMC_HAS_TVM_FFI
 
-#include "plugins/tvm_ffi_module_loader.h"
+#include "runtime/plugins/tvm_ffi/tvm_ffi_module_loader.h"
 
 #include <NvInfer.h>
 #include <cuda_runtime_api.h>
